Split the model-service vlm and detection loader modules into one-class-per-file loaders subpackages with a shared base, keeping each loader.py as a thin aggregator that re-exports the public surface and imports every per-loader module so registration still runs on import and no consumer import path changes.
