+++ docs/docs/project/changelog.md
@@ -84,6 +84,10 @@ The 0.5.0 cycle delivers the architecture-modularization roadmap (`notes/archite
 - The model-service now emits a committed OpenAPI 3.1 contract (`model-service/openapi.json`) for the shapes the server consumes (detection, ontology augmentation, claim extraction, summary synthesis, summarize), and the server generates TypeScript types from it (`pnpm --filter @fovea/server gen:model-service-types` → `src/lib/model-service/contract.ts`). Generation is **ML-free** — `model-service/scripts/gen_contract_spec.py` imports only the Pydantic request/response models (not the FastAPI app) and asserts `cv2`/`torch` never load — so the spec is deterministic and the CI gate does not need the heavy ML install. This is the producer-owned counterpart to the server→frontend OpenAPI generation: the model-service owns the shapes it produces; the server tracks them by generation rather than by hand.
 - `server/src/lib/model-service/contract-assertions.ts` adds **compile-time compatibility assertions** between the generated producer types and the server's hand-written model-service-facing interfaces (`ModelClaimExtractionResponse`, `ModelSynthesisResponse`, `ModelSummarizeResponse`/`Request`, and the detection/ontology request+response expectations). If the model-service drops, renames, or retypes a field the server reads, regenerating the contract makes `tsc` fail with an error whose `offendingKeys` names the exact field — a server↔model-service drift becomes a red build, not a runtime surprise. A new blocking `contract-drift-model-service` CI job (Contracts / Model Service Drift Check) regenerates the spec and the generated types and fails on any diff. This surfaced and corrected one latent annotation drift (`ModelSummarizeResponse.key_frames` was typed `number[]` but the producer sends `KeyFrame` objects, stored verbatim into a JSON column — a type-only fix). `docs/docs/development/cross-service-contracts.md` documents the single workflow for adding or changing a cross-service endpoint.
 
+#### Model-Service Loader Files Split One-Class-Per-File
+
+- The two monolithic model-service loader modules — `vlm/loader.py` (1158 lines, 6 VLM loaders) and `detection/loader.py` (771 lines, 7 detection loaders) — are split into `loaders/` subpackages with one loader per file plus a shared `loaders/base.py` (the config/enums, the abstract base, the architecture-keyed registry, and the factory). Each former `loader.py` is kept as a thin aggregator that re-exports the public surface and imports every per-loader module for its `@registry.register(...)` side effect, so registration still runs on import and no consumer import path changes. Pure reorganization — no loader logic, registry, factory dispatch, or architecture class changed; the registered architectures (9 VLM, 7 detection PyTorch + 3 ONNX), the architecture↔loader invariant test, `mypy`, and the full suite are unchanged.
+
 ### Fixed
 
 #### Reviewers Can Resolve Cross-User Object Names
